+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add built-in colors mimicking Bevy's color palettes (CSS1 'basic', CSS4 'css', Tailwind CSS 'tailwind').
 - Rework widget theming and interactivity to use simpler control scheme. Now you only need `ControlRoot` and `ControlLabel` loadables to set up multi-entity interactive and pseudo-state-sensitive structures.
 - Add font families API for accessing fonts.
+- Implement `Add` for `SceneFile` and `SceneRef`.
 
 ## 0.3.0